moved Spinner to v5/shared
