+++ src/components/frame/Extensions/pages/ExtensionDetailsPage/ExtensionDetailsPage.tsx
@@ -6,7 +6,7 @@ import { Extension } from '@colony/colony-js';
 
 import { useColonyContext, useExtensionData } from '~hooks';
 import ExtensionDetails from './partials/ExtensionDetails';
-import Spinner from '~shared/Extensions/Spinner';
+import Spinner from '~v5/shared/Spinner';
 import ThreeColumns from '~v5/frame/ThreeColumns';
 import Navigation from '~v5/common/Navigation';
 import ImageCarousel from '~common/Extensions/ImageCarousel';
