+++ src/components/frame/Extensions/pages/ReputationPage/ReputationPage.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 
 import Navigation from '~v5/common/Navigation';
 import TwoColumns from '~v5/frame/TwoColumns';
-import Spinner from '~shared/Extensions/Spinner';
+import Spinner from '~v5/shared/Spinner';
 
 const displayName = 'frame.Extensions.pages.ReputationPage';
