+++ src/components/v5/shared/Spinner/index.ts
